pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-equity-momentum"
version = "0.1.0"
description = "Equity momentum rebalance app: paper + live trading on Dhan"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "APScheduler>=3.10",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "pandas>=2.2",
    "psutil>=5.9",
    "pyjwt>=2.8",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "freezegun>=1.4",
    "respx>=0.20",
    "ruff>=0.4",
]

[project.scripts]
emrb-worker = "app.worker.main:main"
emrb-web = "app.web.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"

[tool.ruff]
line-length = 110
target-version = "py311"
